requirements.txt
requests>=2.28.0
pywebview>=4.0.0  # optional — for fullscreen app launching. Falls back to browser if not installed.
